+++ tests/Unit/DSC_VMHyperV.Tests.ps1
@@ -20,7 +20,7 @@ function Invoke-TestSetup
 
     # Import the stub functions.
     Import-Module -Name "$PSScriptRoot/Stubs/Hyper-V.stubs.psm1" -Force
-    Import-Module -Name "$PSScriptRoot/Stubs/FailoverClusters" -Force
+    Import-Module -Name "$PSScriptRoot/Stubs/FailoverClusters.stubs.psm1" -Force
 }
 
 function Invoke-TestCleanup
@@ -60,7 +60,7 @@ try
 
             $mockVmGuid = [System.Guid]::NewGuid().ToString()
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'RunningVM' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'RunningVM' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'RunningVM'
                 $stubVM.HardDrives = @(
@@ -90,7 +90,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'StoppedVM' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'StoppedVM' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'StoppedVM'
                 $stubVM.HardDrives = @(
@@ -120,7 +120,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'PausedVM' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'PausedVM' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'PausedVM'
                 $stubVM.HardDrives = @(
@@ -150,11 +150,13 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'NonexistentVM' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'NonexistentVM' }
+
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'NonexistentVM' -and $ThrowOnEmpty} -MockWith {
                 Write-Error 'VM not found.'
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'DuplicateVM' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'DuplicateVM' } -MockWith {
                 $stubVM1 = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM1.Name = 'DuplicateVM'
                 $stubVM1.HardDrives = @(
@@ -212,7 +214,7 @@ try
                 )
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'Generation1Vhd' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'Generation1Vhd' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'Generation1Vhd'
                 $stubVM.HardDrives = @(
@@ -241,7 +243,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'Generation2VM' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'Generation2VM' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'Generation2VM'
                 $stubVM.HardDrives = @(
@@ -270,7 +272,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'VMWithAutomaticCheckpoints' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'VMWithAutomaticCheckpoints' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'VMWithAutomaticCheckpoints'
                 $stubVM.HardDrives = @(
@@ -300,7 +302,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'VMWithoutAutomaticCheckpoints' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'VMWithoutAutomaticCheckpoints' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'VMWithoutAutomaticCheckpoints'
                 $stubVM.HardDrives = @(
@@ -330,7 +332,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'VMAutomaticCheckpointsUnsupported' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'VMAutomaticCheckpointsUnsupported' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'VMAutomaticCheckpointsUnsupported'
                 $stubVM.HardDrives = @(
@@ -359,7 +361,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'ClusteredExistingVmNotInClusterPresent' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'ClusteredExistingVmNotInClusterPresent' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'ClusteredExistingVmNotInClusterPresent'
                 $stubVM.HardDrives = @(
@@ -389,7 +391,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'ClusteredExistingVmInClusterPresent' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'ClusteredExistingVmInClusterPresent' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'ClusteredExistingVmInClusterPresent'
                 $stubVM.HardDrives = @(
@@ -419,37 +421,7 @@ try
                 return $stubVM
             }
 
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'ClusteredNewVmPresent' } -MockWith {
-                $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
-                $stubVM.Name = 'ClusteredNewVmPresent'
-                $stubVM.HardDrives = @(
-                    $stubVhdxDisk,
-                    $stubVhdDisk
-                )
-                $stubVM.Path = $StubVMConfig.FullPath
-                $stubVM.Generation = 1
-                $stubVM.MemoryStartup = 512MB
-                $stubVM.MemoryMinimum = 128MB
-                $stubVM.MemoryMaximum = 4096MB
-                $stubVM.ProcessorCount = 1
-                $stubVM.ID = $mockVmGuid
-                $stubVM.CPUUsage = 10
-                $stubVM.MemoryAssigned = 512MB
-                $stubVM.Uptime = New-TimeSpan -Hours 12
-                $stubVM.CreationTime = (Get-Date).AddHours(-12)
-                $stubVM.DynamicMemoryEnabled = $true
-                $stubVM.Notes = ''
-                $stubVM.State = 'Running'
-                $stubVM.IsClustered = $false
-                $stubVM.NetworkAdapters = @(
-                    $stubNIC1,
-                    $stubNIC2
-                )
-
-                return $stubVM
-            }
-
-            Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'ClusteredExistingVmInClusterAbsent' } -MockWith {
+            Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'ClusteredExistingVmInClusterAbsent' } -MockWith {
                 $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                 $stubVM.Name = 'ClusteredExistingVmInClusterAbsent'
                 $stubVM.HardDrives = @(
@@ -491,6 +463,7 @@ try
 
             Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'Hyper-V') -and ($ListAvailable -eq $true) } -MockWith { return $true }
             Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'FailoverClusters') -and ($ListAvailable -eq $true) } -MockWith { return $true }
+            Mock -CommandName Get-Command -ParameterFilter {$Name -eq 'Get-Cluster'}
             Mock -CommandName Get-VhdHierarchy -ParameterFilter { $VhdPath.EndsWith('.vhd') } -MockWith {
                 # Return single Vhd chain for .vhds
                 return @($stubVhdDisk.Path)
@@ -538,11 +511,6 @@ try
                     # This test needs to be the last in the Context otherwise all subsequent Get-Module checks will fail
                     Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'Hyper-V') -and ($ListAvailable -eq $true) }
                     { Get-TargetResource -Name 'RunningVM' -VhdPath $stubVhdxDisk.Path } | Should -Throw
-                }
-                It 'throws when Clustering feature is not installed but VM should be clustered' {
-                    # This test needs to be the last in the Context otherwise all subsequent Get-Module checks will fail
-                    Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'FailoverClusters') -and ($ListAvailable -eq $true) } -MockWith { return $true }
-                    { Get-TargetResource -Name 'ClusteredExistingVmInClusterPresent' -VhdPath $stubVhdxDisk.Path -IsClustered $true } | Should -Throw
                 }
             } #end context Validates Get-TargetResource Method
 
@@ -701,18 +669,6 @@ try
                     Test-TargetResource -Name 'VMWithAutomaticCheckpoints' -AutomaticCheckpointsEnabled $false @testParams | Should -Be $false
                 }
 
-                It 'Returns $true when IsClustered is enabled, Ensure is Present and VM is clustered' {
-                    Test-TargetResource -Name ClusteredExistingVmInClusterPresent -IsClustered $true @testParams | Should -Be $true
-                }
-
-                It 'Returns $false when IsClustered is enabled, Ensure is Present and VM is not clustered' {
-                    Test-TargetResource -Name ClusteredExistingVmNotInClusterPresent -IsClustered $true @testParams | Should -Be $false
-                }
-
-                It 'Returns $false when IsClustered is enabled, Ensure is Absent VM is present and VM is clustered' {
-                    Test-TargetResource -Name ClusteredExistingVmInClusterAbsent -Ensure Absent -IsClustered $true @testParams | Should -Be $false
-                }
-
                 It 'Returns $true when EnableGuestService is on and requested "EnableGuestService" = "$true"' {
                     Mock -CommandName Get-VMIntegrationService -MockWith {
                         $guestServiceInterface = [Microsoft.HyperV.PowerShell.VMIntegrationComponent]::CreateTypeInstance()
@@ -731,12 +687,6 @@ try
                     { Test-TargetResource -Name 'RunningVM' @testParams } | Should -Throw
                 }
 
-                It 'throws when Failover Clustering us not installed' {
-                    # This test needs to be the last in the Context otherwise all subsequent Get-Module checks will fail
-                    Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'FailoverClusters') -and ($ListAvailable -eq $true) }
-                    { Test-TargetResource -Name 'RunningVM' -IsClustered $true @testParams} | Should -Throw
-                }
-
             } #end context Validates Test-TargetResource Method
 
             Context 'Validates Set-TargetResource Method' {
@@ -744,7 +694,7 @@ try
                     VhdPath = $stubVhdxDisk.Path
                 }
 
-                Mock -CommandName Get-VM -ParameterFilter { $Name -eq 'NewVM' }
+                Mock -CommandName Get-VMHyperV -ParameterFilter { $VMName -eq 'NewVM' }
                 Mock -CommandName New-VM -MockWith {
                     $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                     $stubVM.Name = 'NewVM'
@@ -956,27 +906,34 @@ try
                     Assert-MockCalled -CommandName Disable-VMIntegrationService -Exactly -Times 1 -Scope It
                 }
 
-                Mock -CommandName Get-ClusterGroup -MockWith {@{
-                    Name = 'VM'
-                    OwnerNode = 'HOST1'
-                    State = 'Online'
-                }}
+                Mock -CommandName Get-ClusterGroup -MockWith {
+                    @{
+                        Name      = 'A VM'
+                        OwnerNode = 'A Cluster Node'
+                    }
+                }
                 Mock -CommandName Remove-ClusterGroup
                 Mock -CommandName Add-ClusterVirtualMachineRole
 
-                It 'Calls Get- and Remove-ClusterGroup if Ensure is Absent and VM exists' {
-                    Set-TargetResource -Name 'ClusteredExistingVmInClusterAbsent' -IsClustered $true -Ensure Absent @testParams
-                    Assert-MockCalled -CommandName Remove-ClusterGroup -Exactly -Times 1 -Scope It
+                It 'Calls Remove-ClusterGroup if Ensure is Absent and VM exists' {
+                    Mock -CommandName Get-Cluster -MockWith {'TheCluster'}
+                    Mock -CommandName Get-Command -ParameterFilter {$Name -eq 'Get-Cluster'} -MockWith {'Get-Cluster'}
+                    Set-TargetResource -Name 'ClusteredExistingVmInClusterAbsent' -Ensure Absent @testParams
                     Assert-MockCalled -CommandName Get-ClusterGroup -Exactly -Times 1 -Scope It
+                    Assert-MockCalled -CommandName Remove-ClusterGroup -Exactly -Times 1 -Scope It
                 }
 
                 It 'Calls Add-ClusterVirtualMachineRole if Ensure is Present and VM exists but is not clustered yet' {
-                    Set-TargetResource -Name 'ClusteredExistingVmNotInClusterPresent' -IsClustered $true @testParams
+                    Mock -CommandName Get-Cluster -MockWith {'TheCluster'}
+                    Mock -CommandName Get-Command -ParameterFilter {$Name -eq 'Get-Cluster'} -MockWith {'Get-Cluster'}
+                    Set-TargetResource -Name 'ClusteredExistingVmNotInClusterPresent' @testParams
                     Assert-MockCalled -CommandName Add-ClusterVirtualMachineRole -Exactly -Times 1 -Scope It
                 }
 
                 It 'Calls Add-ClusterVirtualMachineRole if Ensure is Present and VM does not exist' {
-                    Set-TargetResource -Name 'NonexistentVm' -IsClustered $true @testParams
+                    Mock -CommandName Get-Cluster -MockWith {'TheCluster'}
+                    Mock -CommandName Get-Command -ParameterFilter {$Name -eq 'Get-Cluster'} -MockWith {'Get-Cluster'}
+                    Set-TargetResource -Name 'NonexistentVm' @testParams
                     Assert-MockCalled -CommandName Add-ClusterVirtualMachineRole -Exactly -Times 1 -Scope It
                 }
 
@@ -1083,16 +1040,11 @@ try
                     Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'Hyper-V') -and ($ListAvailable -eq $true) }
                     { Set-TargetResource -Name 'RunningVM' @testParams } | Should -Throw
                 }
-
-                It 'throws when Clustering tools are not installed' {
-                    Mock -CommandName Get-Module -ParameterFilter { ($Name -eq 'FailoverClusters') -and ($ListAvailable -eq $true) }
-                    { Set-TargetResource -Name 'RunningVM' -IsClustered $true @testParams } | Should -Throw
-                }
             } #end context Validates Set-TargetResource Method
 
             Context 'Validates Test-VMSecureBoot Method' {
                 BeforeAll {
-                    Mock -CommandName Get-VM -MockWith {
+                    Mock -CommandName Get-VMHyperV -MockWith {
                         $stubVM = [Microsoft.HyperV.PowerShell.VirtualMachine]::CreateTypeInstance()
                         $stubVM.Name = 'NewVM'
                         $stubVM.Path = $StubVMConfig.FullPath